+++ migrator/runner/all.go
@@ -38,4 +38,5 @@ import (
 	_ "github.com/stackrox/rox/migrator/migrations/m_77_to_m_78_mitre"
 	_ "github.com/stackrox/rox/migrator/migrations/m_78_to_m_79_exclude_openshift_sdn_host_pids_policy"
 	_ "github.com/stackrox/rox/migrator/migrations/m_79_to_m_80_more_openshift_exclusions"
+	_ "github.com/stackrox/rox/migrator/migrations/m_80_to_m_81_rm_nginx_policy"
 )